feat: implement exporting pprof
